Hot or Cold Initial Commit
This is my Hot or Cold project for Thinkful.  It has the basic
functionality down, but there is still a lot of work to do to make it
look better.  Looking to improve the looks, need to fix that you can
still click after you clicked the correct answer, need to remove or
change the 0 row and column.